Refactor: improvments for the unit tests
