+++ tests/mocks/common-configurations.ts
@@ -1,6 +1,6 @@
 import { MalgaFormElements } from '../../src/common/interfaces/form'
 
-export const malgaConfigurations = (isSandbox: boolean) => {
+export function malgaConfigurations(isSandbox: boolean) {
   return {
     apiKey: '17a64c8f-a387-4682-bdd8-d280493715e0',
     clientId: 'd1d2b51a-0446-432a-b055-034518c2660e',
@@ -10,7 +10,7 @@ export const malgaConfigurations = (isSandbox: boolean) => {
   }
 }
 
-export const configureFormSubmissionMock = (eventSubmit?: any) => {
+export function configureFormSubmissionMock(eventSubmit?: any) {
   eventSubmit
     ? (window.HTMLFormElement.prototype.submit = eventSubmit)
     : (window.HTMLFormElement.prototype.submit = () => {})
@@ -27,7 +27,7 @@ export const formValuesMock = {
   cvv: '114',
 }
 
-export const handleFormMock = () => {
+export function handleFormMock() {
   return {
     form: document.createElement('form'),
     holderNameInput: document.createElement('input'),
