+++ tests/mocks/form-dom.ts
@@ -1,0 +1,33 @@
+import {
+  formElementsMock,
+  formValuesMock,
+  handleFormMock,
+} from './common-configurations'
+
+export function generateForm(onSubmit?: any) {
+  const { form, holderNameInput, cvvInput, expirationDateInput, numberInput } =
+    handleFormMock()
+
+  form.setAttribute(formElementsMock.form, '')
+  form.onsubmit = onSubmit
+  form.id = 'form'
+  form.method = 'POST'
+  form.action = '/test'
+
+  holderNameInput.setAttribute(formElementsMock.holderName, '')
+  numberInput.setAttribute(formElementsMock.number, '')
+  cvvInput.setAttribute(formElementsMock.cvv, '')
+  expirationDateInput.setAttribute(formElementsMock.expirationDate, '')
+
+  document.body.appendChild(form)
+  form.appendChild(holderNameInput)
+  form.appendChild(numberInput)
+  form.appendChild(expirationDateInput)
+  form.appendChild(cvvInput)
+
+  const inputs = document.querySelectorAll('input')
+  inputs[0].value = formValuesMock.holderName
+  inputs[1].value = formValuesMock.number
+  inputs[2].value = formValuesMock.expirationDate
+  inputs[3].value = formValuesMock.cvv
+}
